CI passing and some cleanup
