+++ habitat-baselines/habitat_baselines/rl/multi_agent/pop_play_wrappers.py
@@ -1,4 +1,5 @@
 from collections import defaultdict
+from dataclasses import dataclass
 from typing import Any, Callable, Dict, List, Optional
 
 import gym.spaces as spaces
@@ -12,12 +13,68 @@ from habitat_baselines.rl.multi_agent.utils import (
     add_agent_prefix,
     update_dict_with_agent_prefix,
 )
-from habitat_baselines.rl.ppo.policy import (
-    MultiAgentPolicyActionData,
-    Policy,
-    PolicyActionData,
-)
+from habitat_baselines.rl.ppo.policy import Policy, PolicyActionData
 from habitat_baselines.rl.ppo.updater import Updater
+
+
+@dataclass
+class MultiAgentPolicyActionData(PolicyActionData):
+    """
+    Information returned from the `Policy.act` method representing the
+    information from multiple agent's action. This class is needed to store
+    actions of multiple agents together
+    :property length_actions: List containing, for every agent, the size of their action space.
+    :property length_rnn_hidden_states: List containing for every agent the dimensionality of the rnn hidden state.
+    :property num_agents: The number of agents represented in this PolicyActionData
+    """
+
+    rnn_hidden_states: torch.Tensor
+    actions: Optional[torch.Tensor] = None
+    values: Optional[torch.Tensor] = None
+    action_log_probs: Optional[torch.Tensor] = None
+    take_actions: Optional[torch.Tensor] = None
+    policy_info: Optional[List[Dict[str, Any]]] = None
+    should_inserts: Optional[np.ndarray] = None
+
+    # Indices
+    length_rnn_hidden_states: Optional[torch.Tensor] = None
+    length_actions: Optional[torch.Tensor] = None
+    num_agents: Optional[int] = 1
+
+    def _unpack(self, tensor_to_unpack, unpack_lengths=None):
+        """
+        Splits the tensor tensor_to_unpack in the last dimension in the last dimension
+        according to unpack lengths, so that the ith tensor will have unpack_lengths[i]
+        in the last dimension. If unpack_lenghts is None, splits tensor_to_unpack evenly
+        according to self.num_agents.
+        :property tensor_to_unpack: The tensor we want to split into different chunks
+        :unpack_lengths: List of integers indicating the sizes to unpack, or None if we want to unpack evenly
+        """
+
+        if unpack_lengths is None:
+            unpack_lengths = [
+                int(tensor_to_unpack.shape[-1] / self.num_agents)
+            ] * self.num_agents
+
+        return torch.split(tensor_to_unpack, unpack_lengths, dim=-1)
+
+    def unpack(self):
+        """
+        Returns attributes of the policy unpacked per agent
+        """
+        return {
+            "next_recurrent_hidden_states": self._unpack(
+                self.rnn_hidden_states, self.length_rnn_hidden_states
+            ),
+            "actions": self._unpack(self.actions, self.length_actions),
+            "value_preds": self._unpack(self.values),
+            "action_log_probs": self._unpack(self.action_log_probs),
+            "take_actions": self._unpack(self.take_actions),
+            # This is numpy array and must be split differently.
+            "should_inserts": np.split(
+                self.should_inserts, self.num_agents, axis=-1
+            ),
+        }
 
 
 class MultiPolicy(Policy):
