+++ habitat-lab/habitat/core/env.py
@@ -143,11 +143,6 @@ class Env:
             for k, v in self._config.environment.iterator_options.items()
         }
         iter_option_dict["seed"] = self._config.seed
-        if (
-            "overfit" in self._config.simulator
-            and self._config.simulator["overfit"]
-        ):
-            iter_option_dict["num_episode_sample"] = 1
 
         self._episode_iterator = self._dataset.get_episode_iterator(
             **iter_option_dict
